Add more instances of explicit typing; use more np.floating for parameter types

+++ pyuvdata/uvdata/uvfits.py
@@ -450,15 +450,15 @@ class UVFITS(UVData):
                 self.spw_array = uvutils._fits_gethduaxis(vis_hdu, 5).astype(np.int) - 1
 
                 # the axis number for phase center depends on if the spw exists
-                self.phase_center_ra_degrees = np.float(vis_hdr.pop("CRVAL6"))
-                self.phase_center_dec_degrees = np.float(vis_hdr.pop("CRVAL7"))
+                self.phase_center_ra_degrees = np.float64(vis_hdr.pop("CRVAL6"))
+                self.phase_center_dec_degrees = np.float64(vis_hdr.pop("CRVAL7"))
             else:
                 self.Nspws = 1
                 self.spw_array = np.array([np.int(0)])
 
                 # the axis number for phase center depends on if the spw exists
-                self.phase_center_ra_degrees = np.float(vis_hdr.pop("CRVAL5"))
-                self.phase_center_dec_degrees = np.float(vis_hdr.pop("CRVAL6"))
+                self.phase_center_ra_degrees = np.float64(vis_hdr.pop("CRVAL5"))
+                self.phase_center_dec_degrees = np.float64(vis_hdr.pop("CRVAL6"))
 
             # get shapes
             self.Npols = vis_hdr.pop("NAXIS3")
@@ -783,11 +783,11 @@ class UVFITS(UVData):
                 ]
 
             start_freq_array = np.reshape(np.array(start_freq_array), (1, -1)).astype(
-                np.float
+                np.float64
             )
 
             delta_freq_array = np.reshape(np.array(delta_freq_array), (1, -1)).astype(
-                np.float
+                np.float64
             )
 
             # We've constructed a couple of lists with relevant values, now time to
@@ -814,7 +814,7 @@ class UVFITS(UVData):
             # other exciting things...
             ref_freq = start_freq_array[0, 0]
         else:
-            delta_freq_array = np.array([[self.channel_width]]).astype(np.float)
+            delta_freq_array = np.array([[self.channel_width]]).astype(np.float64)
             ref_freq = self.freq_array[0, 0]
 
         if self.Npols > 1:
